added review suggestions
